Update cc controller

+++ src/main/java/edu/northeastern/cs5500/delivery/model/CreditCard.java
@@ -10,9 +10,9 @@ import org.bson.types.ObjectId;
  * of userName and a LocalDate to store expirationDate.
  */
 @Data
-public class CreditCard implements Model{
+public class CreditCard implements Model {
     private ObjectId id;
-    private Integer cardNumber;
+    private Long cardNumber;
     private String userName;
     private LocalDate expirationDate;
